First Commit v1
This is the first version. New Job is working as well as autoValues
after much of the hassles. Show jobs is working and inserts are working
in this build. Dates are working when they are string too.  JobCreation
Collection is working. The sofware is in working condition for now
before i mess up somewhere.